+++ package.js
@@ -5,21 +5,21 @@ Package.describe({
 });
 
 Package.onUse(function (api) {
-  api.versionsFrom("METEOR-CORE@0.9.0-atm");
+  api.versionsFrom("METEOR@0.9.0");
 
   api.use(['jquery', 'stylus', 'coffeescript'], 'client');
   api.use(['ui', 'templating'], 'client');
 
-  api.use("mrt:bootstrap-3@3.2.0-1", 'client');
+  api.use("mizzao:bootstrap-3@3.2.0", 'client');
 
-  api.add_files('templates.html', 'client');
-  api.add_files('tutorial.styl', 'client');
+  api.addFiles('templates.html', 'client');
+  api.addFiles('tutorial.styl', 'client');
 
-  api.add_files('eventEmitter.coffee', 'client');
-  api.add_files('drags.js', 'client');
+  api.addFiles('eventEmitter.coffee', 'client');
+  api.addFiles('drags.js', 'client');
 
-  api.add_files('tutorial.coffee', 'client');
-  api.add_files('helpers.coffee', 'client');
+  api.addFiles('tutorial.coffee', 'client');
+  api.addFiles('helpers.coffee', 'client');
 
   api.export('EventEmitter', 'client');
 });